Fully split gamepadshift from gamepad
