+++ ports/atmel-samd/supervisor/port.c
@@ -71,6 +71,9 @@
 #if CIRCUITPY_GAMEPAD
 #include "shared-module/gamepad/__init__.h"
 #endif
+#if CIRCUITPY_GAMEPADSHIFT
+#include "shared-module/gamepadshift/__init__.h"
+#endif
 #include "shared-module/_pew/PewPew.h"
 
 extern volatile bool mp_msc_enabled;
@@ -228,6 +231,9 @@ void reset_port(void) {
 
 #if CIRCUITPY_GAMEPAD
     gamepad_reset();
+#endif
+#if CIRCUITPY_GAMEPADSHIFT
+    gamepadshift_reset();
 #endif
 #if CIRCUITPY_PEW
     pew_reset();
