+++ shared-bindings/digitalio/DigitalInOut.c
@@ -374,3 +374,14 @@ const mp_obj_type_t digitalio_digitalinout_type = {
     .make_new = digitalio_digitalinout_make_new,
     .locals_dict = (mp_obj_t)&digitalio_digitalinout_locals_dict,
 };
+
+// Helper for validating digitalio.DigitalInOut arguments
+digitalio_digitalinout_obj_t *assert_digitalinout(mp_obj_t obj) {
+    if (!MP_OBJ_IS_TYPE(obj, &digitalio_digitalinout_type)) {
+        mp_raise_TypeError(translate("argument num/types mismatch"));
+    }
+    digitalio_digitalinout_obj_t *pin = MP_OBJ_TO_PTR(obj);
+    raise_error_if_deinited(
+        common_hal_digitalio_digitalinout_deinited(pin));
+    return pin;
+}
